refactor: 보드 컬럼 명 수정 및 의존성 주입으로 dao 생성
